[Utils cleanup] Move unsqueeze utilities to the utils module
