+++ backpack/core/derivatives/batchnorm1d.py
@@ -2,7 +2,7 @@ from torch.nn import BatchNorm1d
 
 from ...utils.utils import einsum
 from .basederivatives import BaseDerivatives
-from .utils import jmp_unsqueeze_if_missing_dim
+from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
 
 
 class BatchNorm1dDerivatives(BaseDerivatives):
