+++ backpack/core/derivatives/conv2d.py
@@ -6,7 +6,7 @@ from ...core.layers import Conv2dConcat
 from ...utils import conv as convUtils
 from ...utils.utils import einsum
 from .basederivatives import BaseDerivatives
-from .utils import jmp_unsqueeze_if_missing_dim
+from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
 
 
 class Conv2DDerivatives(BaseDerivatives):
