+++ backpack/core/derivatives/maxpool2d.py
@@ -4,7 +4,7 @@ from torch.nn.functional import max_pool2d
 
 from ...utils import conv as convUtils
 from .basederivatives import BaseDerivatives
-from .utils import jmp_unsqueeze_if_missing_dim
+from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
 
 
 class MaxPool2DDerivatives(BaseDerivatives):
