+++ backpack/core/derivatives/zeropad2d.py
@@ -3,7 +3,7 @@ from torch.nn.functional import pad
 
 from ...utils.utils import einsum
 from .basederivatives import BaseDerivatives
-from .utils import jmp_unsqueeze_if_missing_dim
+from backpack.utils.unsqueeze import jmp_unsqueeze_if_missing_dim
 
 
 class ZeroPad2dDerivatives(BaseDerivatives):
